builds

+++ src/ir/possible-contents.h
@@ -571,7 +571,7 @@ template<> struct hash<wasm::ConeReadLocation> {
   size_t operator()(const wasm::ConeReadLocation& loc) const {
     return std::hash<
       std::pair<wasm::HeapType, std::pair<wasm::Index, wasm::Index>>>{}(
-      {size_t(loc.type), {loc.depth, loc.index}});
+      {loc.type, {loc.depth, loc.index}});
   }
 };
